Allswap_djangoREST/backend/allswap/products/api/views.py
from rest_framework.generics import ListAPIView, RetrieveAPIView
from products.models import Product
from .serializers import ProductSerializer

class ProductListView(ListAPIView):
    queryset = Product.objects.all()
    serializer_class = ProductSerializer


class ProductDetailView(RetrieveAPIView):
    queryset = Product.objects.all()
    serializer_class = ProductSerializer